Make the subject selector dialog functional
The subject selector dialog is now functional. It gets all the subjects,
displays them in a list and "returns" the clicked one. The add task
screen listens for the result and displays it, or the resource string if
no subject has yet been selected.

Note: Since there is currently no way of inserting a subject into the
database, this has not been tested yet. It will most likely be enhanced
when a way to add subjects was added.