Add basic view task screen
This adds a basic view task screen. It will show you all the
informations, such as the title, the due and reminder start date, the
subject name and the description. You can now also click on one task in
the list view on the home screen and it will bring you to the view task
screen.