Fix crash when selecting "No reminder"
This patch fixes a crash when selecting the "No reminder" option. The
crash was caused by us setting the reminder start date to null when
selecting this option. Then we would non null assert it later when
adding the task. Now we just set it to the due date if no reminder has
been selected.