Add basic subject settings
This adds basic subjects settings, accessible through the settings icon
in the app bar on the home screen. For now, it only lists the subjects,
but this will be expanded for creating and editing. We use
https://github.com/alorma/Compose-Settings for making a
preferences-like layout in compose, through which you can get to the
actual subject settings. In that screen (SettingsScreen) we can then
add more settings as the app grows.