Add feature to add new subjects
By using the floating action button in the "Edit Subjects" setting, you
can now add a new subject. You will have to provide a name, an
abbreviation and a color. A color will be chosen at random at the
start, but you can modify it by clicking on the box that represents the
color.