Add feature to delete subjects
This adds a feature to delete subjects. This is way more complicated
than just removing the row from the database, since tasks that refer to
it's id would suddenly point to a non-existent subject. To deal with
this, we ask the user to which subject we should "move" (aka change the
field in the database) the task to. Then, when the user confirms, we
first move the tasks and then delete the subject. We should expand on
this later, by for example giving the user the option to choose for
each task what should happen with it (for example delete, move, ...).