Change the way tasks are displayed on the home screen
Previously, the tasks were grouped by their subject. Now, we first
make groups of the due dates, and then in those groups are the tasks
grouped by their subject.
Thus the hierarchy is now:

Due Date e.g. November 23, 2021
    My Subject
        Task 1
        Task 2
        ...
    ...
...

Note: This is very early, which means that it's not beautiful nor
functional, so for example you can't click on the tasks.