Add support for adding multiple tasks at once
You can now insert multiple tasks at once by using the button in the
app bar on the add task screen. If you click on it, it adds another
item with a title and a description to the bottom of the screen.
The information above (due date, reminder and subject) are transferred
to each subject, while the title and the description come from the
individual text fields.