Allow deletion of subjects without conflicts
If there are no conflicts (so no tasks refer to that subject) you can
now delete the subject too. If you click on the trash icon, a
confirmation dialog will pop up and through that you can delete the
subject. This patch also renames the screen (previously
EditSubjectsScreen) to SubjectsScreen, since that makes it more clear
what it does.