Delete tasks whose due date passed
Now, if the due date of a task passed (so if the due date is yesterday
or before), the home screen's view model will delete that task without
adding it to the list. It will now also show the completed tasks at the
bottom of their subject group. Also, if all tasks of a subject group
are completed, the subject group itself will be shown at the bottom of
the date group.