Add basic notification worker
This adds a basic worker implementation. This worker is later going to
be used to create notifications for tasks that need to be reminded for.
The MainActivity calls ensureStarted which checks if the worker is
enqueued or running and if it is not, it enqueues it. The worker
schedules itself to run at 12 AM for now, which should however be a
setting.