Properly implement notification settings
Previously, the notification settings were not completely filled out
yet, as the preference tiles needed to be tested. Now, they consist of
a setting for the notification target hour (so the hour at which the
user wants to be reminded) and the notification priority.
This patch also changes a few minor things with how the preferences
work. They now need a map of an int and a string, of which the string
is displayed and the int is used to store the value in the preferences.