Fix a bug, where date groups appeared multiple times
Previously, in HomeScreenViewModel::addDate, we would not break out of
the loop when we found the correct spot. This meant that we would
insert a date multiple times if there were dates after the first hit.
Since the dates array is an index for the groups map, the home screen
composable would then render the same date group multiple times.