Format the description of tasks and exams
The description of tasks and exams now supports formatting. The view
task screen now has a custom field for the description in which the
things you typed in are being formatted and displayed in a nice way.

Currently supported formats:
    - Unordered lists (aka. bullet point lists)
    - Ordered lists (1., 2., ...)
    - Links (format: [<link>](<text to be displayed>), will result in a
    underlined clickable piece of text which, on click, opens the link)

You can put as many of these formats in the description as you want.
They will be formatted at runtime and only displayed in the dedicated
field in the view task screen. In the add- and edit screens you will
still see the text form to be able to properly edit it.