feat: Option to save a draft when adding
If the user presses the back button in the add task screen and they
have unsaved changes, a dialog will appear warning the user that their
changes will be lost and giving them the option to save their changes
as a draft. These saved changes will then reappear when entering the
add task screen again. If the user decides to not save a draft,
previously saved data will be removed as well.